requirements.txt
requests
python-dotenv
notebook
pandas
langchain-openai
langchain-core